src/requirements.txt
psycopg2-binary==2.9.9
cfnresponse==1.1.2